# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Dhikr & Dua App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quran/log:
    get:
      operationId: getQuranLog
      tags: [quran]
      summary: Get Quran reading log
      parameters:
        - name: sessionId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Quran reading log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuranLog"
    post:
      operationId: saveQuranLog
      tags: [quran]
      summary: Save a Quran reading session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveQuranLogRequest"
      responses:
        "200":
          description: Quran log saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuranLog"
  /quran/memorisation:
    get:
      operationId: getMemorisation
      tags: [quran]
      summary: Get memorisation log
      parameters:
        - name: sessionId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Memorisation log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemorisationLog"
    post:
      operationId: saveMemorisation
      tags: [quran]
      summary: Save memorisation progress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveMemorisationRequest"
      responses:
        "200":
          description: Memorisation saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemorisationLog"
...
  /stories:
    get:
      operationId: getStories
      tags: [stories]
      summary: Get stories in a category
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of stories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Story"
  /stories/{id}:
    get:
      operationId: getStory
      tags: [stories]
      summary: Get a single story
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Story detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
components:
...
  schemas:
...
    QuranLog:
      type: object
      properties:
        sessionId:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/QuranEntry"
        totalPages:
          type: integer
        totalJuz:
          type: number
        streak:
          type: integer
        lastRead:
          type: string
      required: [sessionId, entries, totalPages, totalJuz, streak]
    QuranEntry:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
        surahFrom:
          type: integer
        ayahFrom:
          type: integer
        surahTo:
          type: integer
        ayahTo:
          type: integer
        pages:
          type: integer
        duration:
          type: integer
        notes:
          type: string
      required: [id, date, surahFrom, ayahFrom, surahTo, ayahTo, pages]
    SaveQuranLogRequest:
      type: object
      properties:
        sessionId:
          type: string
        surahFrom:
          type: integer
        ayahFrom:
          type: integer
        surahTo:
          type: integer
        ayahTo:
          type: integer
        pages:
          type: integer
        duration:
          type: integer
        notes:
          type: string
      required: [sessionId, surahFrom, ayahFrom, surahTo, ayahTo, pages]
    MemorisationLog:
      type: object
      properties:
        sessionId:
          type: string
        memorisedAyahs:
          type: array
          items:
            $ref: "#/components/schemas/MemorisedAyah"
        totalAyahs:
          type: integer
        currentSurah:
          type: integer
        currentAyah:
          type: integer
        lastReviewed:
          type: string
      required: [sessionId, memorisedAyahs, totalAyahs]
    MemorisedAyah:
      type: object
      properties:
        surah:
          type: integer
        ayah:
          type: integer
        surahName:
          type: string
        strength:
          type: string
          enum: [new, learning, review, strong]
        lastReviewed:
          type: string
      required: [surah, ayah, surahName, strength]
    SaveMemorisationRequest:
      type: object
      properties:
        sessionId:
          type: string
        surah:
          type: integer
        ayah:
          type: integer
        surahName:
          type: string
        strength:
          type: string
          enum: [new, learning, review, strong]
      required: [sessionId, surah, ayah, surahName, strength]
...
    Story:
      type: object
      properties:
        id:
          type: string
        categoryId:
          type: string
        title:
          type: string
        subtitle:
          type: string
        content:
          type: array
          items:
            $ref: "#/components/schemas/StorySection"
        arabicTitle:
          type: string
        readTime:
          type: integer
        tags:
          type: array
          items:
            type: string
      required: [id, categoryId, title, content, readTime]
    StorySection:
      type: object
      properties:
        type:
          type: string
          enum: [paragraph, hadith, verse, highlight]
        text:
          type: string
        arabic:
          type: string
        reference:
          type: string
      required: [type, text]
...