# lib/api-zod/src/generated/types/quranEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dhikr & Dua App API
 * OpenAPI spec version: 0.2.0
 */

export interface QuranEntry {
  id: string;
  date: string;
  surahFrom: number;
  ayahFrom: number;
  surahTo: number;
  ayahTo: number;
  pages: number;
  duration?: number;
  notes?: string;
}