# lib/api-zod/src/generated/types/quranLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dhikr & Dua App API
 * OpenAPI spec version: 0.2.0
 */
import type { QuranEntry } from "./quranEntry";

export interface QuranLog {
  sessionId: string;
  entries: QuranEntry[];
  totalPages: number;
  totalJuz: number;
  streak: number;
  lastRead?: string;
}