# lib/api-zod/src/generated/types/storySection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dhikr & Dua App API
 * OpenAPI spec version: 0.2.0
 */
import type { StorySectionType } from "./storySectionType";

export interface StorySection {
  type: StorySectionType;
  text: string;
  arabic?: string;
  reference?: string;
}